add conditional exits if apply or plan fails

+++ deployment/staging/s3_deploy/upload_s3/replace_files.sh
@@ -13,7 +13,24 @@ find ../../../../dist -type f > ./dist_file_paths.txt
 echo -e "\n" >> ./vars.tfvars
 awk -f ./dist_paths_to_list.awk ./dist_file_paths.txt >> ./vars.tfvars
 
+terraform plan -var-file=./vars.tfvars
+STATUS=$?
+
+if [ $STATUS -gt 0 ]
+then
+    echo "plan failed"
+    exit $STATUS
+fi
+
 terraform destroy -auto-approve -var-file=./vars.tfvars
 terraform apply -var-file=./vars.tfvars -auto-approve
 
-exit
+STATUS=$?
+
+if [ $STATUS -gt 0 ]
+then
+    echo "apply failed"
+    exit $STATUS
+fi
+
+exit 0
